feat(STEP-1): clone repository and create structure
