+++ src/components/Editor.tsx
@@ -1,0 +1,38 @@
+export default function Editor() {
+  return (
+    <>
+
+
+      <div className="editor-page">
+        <div className="container page">
+          <div className="row">
+            <div className="col-md-10 offset-md-1 col-xs-12">
+              <form>
+                <fieldset>
+                  <fieldset className="form-group">
+                    <input type="text" className="form-control form-control-lg" placeholder="Article Title" />
+                  </fieldset>
+                  <fieldset className="form-group">
+                    <input type="text" className="form-control" placeholder="What's this article about?" />
+                  </fieldset>
+                  <fieldset className="form-group">
+                    <textarea className="form-control" rows={8} placeholder="Write your article (in markdown)" />
+                  </fieldset>
+                  <fieldset className="form-group">
+                    <input type="text" className="form-control" placeholder="Enter tags" />
+                    <div className="tag-list" />
+                  </fieldset>
+                  <button className="btn btn-lg pull-xs-right btn-primary" type="button">
+                    Publish Article
+                  </button>
+                </fieldset>
+              </form>
+            </div>
+          </div>
+        </div>
+      </div>
+
+     
+    </>
+  );
+}
