+++ src/components/Logout.tsx
@@ -1,0 +1,8 @@
+export default function Logout() {
+  return (
+    <>
+
+      
+    </>
+  );
+}
