+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export {Navbar} from './Navbar';
+export {Footer} from './Footer';
