+++ tsconfig.json
@@ -15,7 +15,10 @@
     "isolatedModules": true,
     "noEmit": true,
     "jsx": "react-jsx",
-    "baseUrl": "src"
+    "baseUrl": "src",
   },
-  "include": ["src"]
+  "include": ["src"],
+  "exclude": [
+        "node_modules"
+    ],
 }
